+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * added `encoding` to `read_lines`
 * added `none_if_false`
 * fixed import mishap in `satella.coding.transforms` with `hashables_to_int`
+* fixed `read_in_file` if file does not exist and default is not set